3.5.12

+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/RelationshipCountsProcessor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2002-2020 "Neo4j,"
+ * Copyright (c) 2002-2019 "Neo4j,"
  * Neo4j Sweden AB [http://neo4j.com]
  *
  * This file is part of Neo4j.
@@ -68,8 +68,8 @@ public class RelationshipCountsProcessor implements RecordProcessor<Relationship
 
     static long calculateMemoryUsage( int highLabelId, int highRelationshipTypeId )
     {
-        long labels = highLabelId + 1;
-        long types = highRelationshipTypeId + 1;
+        int labels = highLabelId + 1;
+        int types = highRelationshipTypeId + 1;
         long labelsCountsUsage = labels * types * SIDES * Long.BYTES;
         long wildcardCountsUsage = types * Long.BYTES;
         return labelsCountsUsage + wildcardCountsUsage;
